modularized functions and removed them from the component to store inside utils folder
